node_config: allow more connection info options using strings or dicts
Handle connection_strings in the traditional libpq "key=value key=value ..."
format and the new style postgres:// and postgresql:// uris as well as
allowing `key: value` pairs of dicts.

Also make sure we allow all postgresql connection keywords in any input
format but force the connection to be a replication connection to the
specified host using the "replication" pseudo database.