configuration: rename and restructure some configuration keys
Rename `backup_clusters` to `backup_sites` as `site` is the concept we use
internally and also used in some parts of the README.  Also move the node
configuration under a `nodes` key in site configuration and make it a list
instead of a dict, requiring all parameters of a connection to be defined
inside the connection info / string.

Also refactored and restructured some other code related to using these
keys.