pghoard: support encryption and integrity protection of the backup files
Backed up files can be encrypted at rest. Encryption is configured via
providing RSA public key (for backup / encryption) and/or RSA private
key (for restore / decryption) in pghoad configuration file. If enabled,
each segment is encrypted and authenticated with fresh key material. Key
material is encrypted with RSA keys and stored in the beginning of each
encoded file segment. Message authentication code is appended to the
end of the actual content data.