webserver: use the new file listing api from transferagent
Implement a file listing method in TransferAgent and use it to list
basebackups so we can list remote basebackups as well as local ones.

Also updated Restore's HTTPRestore to subclass ObjectStore so we use the
same interface for listing and showing basebackups whether we use direct
Object Storage access or the internal HTTP API.

And while touching this code, let's change the HTTP path for basebackups
from `basebackups` to `basebackup` which is what we use everywhere else.