archive_sync: try to sync all files since last basebackup
Previously we looked up WAL files from newest to oldest and stopped lookups
when we found a WAL file that's already archived.  This had a race condition
when PG was running and archived more WAL files before we had a chance to
run the sync and archive old ones.

Update the logic to look up the latest basebackup and the first WAL segment
it needs and try to archive all segments between it and the currently open
one.