google: check if bucket exists before trying to create it
Look up the bucket if it already exists and only try to create the bucket in
case it doesn't.  Note that we can't just always try to unconditionally
create the bucket as Google imposes a strict rate limit on bucket creation
operations, even if it doesn't result in a new bucket.