object delete: return and handle a known exception type for non-existent keys
Google storage driver passed Google's 404 HttpError back to the caller if we
tried to delete a non-existent key which caused an ugly traceback to be
printed and didn't break our "delete wal before" loop.  S3 returned False
which resulted in debug log and breaking the loop.

Updated both S3 and Google drivers to raise FileNotFoundFromStorageError
instead which we already use for GETs of non-existent files.  The error now
causes logging and break of delete loop.

We ran into this while trying to delete WAL segments of previous timelines,
will fix timeline traversal in another commit.