delete_remote_wal_before: handle timeline traversal
If we fail to delete a WAL segment on a given timeline try deleting the same
segment number on a previous timeline, but flag that timeline as "invalid"
until we're able to delete at least one segment on it.

This way we know when to stop timeline traversal as there is no point in
trying to delete more WAL segments on timelines that didn't have any to
start with.  Also never try to step back beyond timeline 1.