basebackup pruning: fix file paths in basebackup pruning
Object storages' list_path() operation is supposed to return full path
(relative to the object store of course) to each entry in path, but in the
case of local storage we only returned the base file name.  In cloud
storages we returned full path and when we later tried to delete excess
basebackups it'd fail as we again prepended prefix and "basebackups/" to the
key.  Fix this by returning full key in local storage and stripping paths in
basebackup listing, as the list is supposed to include basebackups' names
and not full filenames.