webserver: refactor for uniform paths and response generation
Refactored webserver to use common functions for parsing and handling
requests as well as generation responses for them.  We also make sure to
always generate a response even if there was an internal timeout or
something like that.

Also updated paths to include object types in them and dropped the `archive`
tag from paths used in uploads.

This commit also drops the ability to download basebackups over HTTP from a
local pghoard server.  The way it was implemented was incompatible with
object storages and because of that it wasn't very useful.  We'll bring it
back later with proper object storage support.