tests: use flake8 with max-line-len 125 instead of pep8
flake8 performs the same checks as pep8 plus some additional ones similar to
Pylint's checks but also catches things that Pylint doesn't, and it's a lot
faster.  So let's replace pep8 checks with it.  We still run pylint as well.

Also set max line length to 125 characters which is what you can see on a
GitHub diff review.