webserver: prefetch files on restore
Prefetch files when performing archive recovery using pghoard's restore
command against the local pghoard webserver.  The default is to start up to
'lower of Compression or Transfer Agent threads minus one' prefetch
operations so we can perform all operations (original request + prefetch)
fully in parallel without having to wait for threads to start.

We maintain a negative cache for "file not found" prefetch failures to avoid
trying to prefetch the same files multiple times, but the cache isn't used
for explicit download requests as the segments may show up in the archive
after the first failures.

This speeds up archive recovery a bit in default configurations, but it may
make sense to increase thread counts beyond the current defaults (5) if the
store is slow or has high latency.