logging: automatically adjust log format if we are running under systemd
If we're running under systemd the output goes to journal which includes
timestamps.  Prefer our syslog log format in case the log is going to
journal, it doesn't include timestamps or use tabs, making it easier to
process.  It didn't include thread name previously, this commit adds it to
the format.

Also output a warning if we're running under systemd but python-systemd
isn't installed.  When the default unit file is used systemd is waiting for
PGHoard to notify it when it's up and running, but without python-systemd
the notification never happens and systemd won't know we have started.