python 3.3: fix issues with hex codec and tarfile
In Python 3.3's tarfile there's a bug - which has been fixed in later
versions - that if there's anything set as the .name attribute of a
fileobject it gets to be the name.  In this particular case Python sets it
to the integer 4.  Later versions of Python check that it has to be either a
bytes object or a string in order for it to use it.

Also changed `hex` codec name to `hex_codec` since it's defined in both
Python 3.3 and 3.4, unlike just plain `hex` codec.