s3: cleanups and more test coverage for multipart uploads
Make sure we read the source file to end instead of relying on chunk count
calculation.  This should not make a difference in practice, but makes sure
we'll eventually error out if the file changes under us.

Also improved log and exception messages and added test coverage for
temporary failures in chunk uploads.