config: add functions for reading and parsing config and finding a site in it
Add functions in pghoard.config to read a JSON configuration file and check
the backup sites in it.  This makes it possible to use various commands with
just a --config flag when they define a single site which is the typical use
case.

This commit also makes sure the various places that read configs work in the
same way.