compressor: don't raise an exception in the main loop if an event fails
Previously we'd cause the compressor thread to die with an exception when
event handling failed.  That would probably end up with a stuck pghoard at
some point when all compressor threads were dead.

Instead of that, let's push an error in the event's callback queue (if any)
and log the exception and keep going.