local-tar backup: manually iterate over files to backup
This allows us to skip temporary files everywhere and to avoid raising
errors if files disappear under us.  We don't care if files in various
directories are deleted, as PostgreSQL may do it when objects are dropped or
truncated.  The WAL we require lists all the operations that may cause
filesystem changes.

The manual file collection code wil also be used by the upcoming parallel
backup mechanism.