basebackups: store backup end time and segment in metadata in local-tar mode
Basebackup `end-time` and `end-wal-segment` are now stored in metadata and
used for PITR when `local-tar` basebackups are used.

When performing recovery to a given point the point must be after the end of
the backup, so try to store such a time in the metadata.  The pg_basebackup
methods don't currently store it as there's no way to get it over a
replication protocol.  We'll probably require a non-replication connection
to the db in the future for all basebackup types.