tests: Fix webserver unittest failure by using a different WAL segment
The test_handle_site in test_basebackup created the same WAL file
and if it was run before test_get_encrypted_archived_file the test
would fail since it would return the pre-existing data.

The reason the data passes the webserver code's prefetch correctness
checks is that it really had the correct WAL file header.