pghoard: Add config validation CL parameter.
Addresses #188. This adds syntactic checking for pghoard config (i.e., it will
check that your S3 credentials are properly filled-in, but not if the given
bucket can be accessed using the given keys). This is useful for tools like
Ansible which can validate files before copying, or for reloader scripts to check
that configuration is valid before reloading pghoard after a config change.