Remove excessive bytes slicing from decryption
Perf testing shows that restoring backups is slow because a lot of time
is spent doing memmove. The decryption code does a whole lot of slicing
and joins for bytes and contributes to the overhead. Remove most of the
slicing and combining of byte arrays.

This implementation also fixes seek to end-of-file handling, which
previously read and decrypted full file contents into memory and then
threw the results away. Seeking to eof is common for getting the size of
a file, which we can get practically as a zero op.

Note that this implementation only works as long as AES-CTR encryption
is being used since it makes assumptions regarding unencrypted data
that may not hold for other ciphers / cipher modes.