webserver: Do proper concurrent WAL/history downloads
Earlier logic tried to prefetch WAL/history files that are expected to
be requested by PostgreSQL but because it waited for the prefetch ops
to complete before serving the original request it wasn't very
effective. Also, due to GIL the processing often ended up being CPU
bound.

To speed up the processing, make all file fetch operations run in the
background as identical jobs and just poll a queue of completed
operations to finalize the background jobs and in case one of them is
the operation that is currently expected to finish, fulfil the request.
New prefetch operations are always scheduled even if the current
request can be fulfilled without downloading anything so that there is
always a backlog of prefetch operations.

In addition to improving the prefetch logic, the file download,
decrypt and decompress operation was changed to be executed in a
separate process so that full CPU capacity of the host can be utilized.
To avoid having a lot of background processes running when initial WAL
playback isn't ongoing, a custom process management logic is used and
idle processes are killed after 10 minutes.

The new implementation was tested with 21 gigabytes of non-trivial (100%
non-compressible inserts) WAL files on a 16 core host with max disk
write speed of about 400 megabytes per second. Processing was restricted
by PostgreSQL not requesting/handling files fast enough or disk I/O,
slightly varying during the run. For the 1310 file requests done during
the run, 1306 were served immediately and the remaining 4 had total
delay of less than 3 seconds. Total restoration time for the run was
220 seconds.

The old version restored the 21 gigabytes of WAL in 340 seconds so the
new implementation is "only" 35% faster but the bottleneck has been
moved completely to PG and I/O side and in case of simple WAL files
faster I/O will make the new version execute better in comparison
and if PG is ever changed to support some kind of concurrency the new
implementation will be able to serve files fast enough.