webserver: Fix handling of timed out downloads
If a download has taken longer than it should, a new download operation
for the same file is started. Previously this overwrote the expected
name of the temporary file that shall be created by the download. This
caused problems if the earlier download did eventually finish and it
also finished before the newly scheduled download. The result handler
was called for the first download as well but instead of using the
original temp file name it used the second download's temp file name,
which was still in progress and the file was empty.

The operation was retried for 30 seconds and future downloads might
have succeeded but depending on timing conditions there could've been
multiple failures and the download failed permanently, causing WAL
restoration to fail.

Currently missing unittest for exercising the fix because there's no
easy way to mock behavior in the remote process and more isolated
test cannot be done without major refactoring.