basebackup: Ensure each tar can be extracted individually
Previously the directory create operations were only part of the tar
file that contained first file/dir under the directory. This resulted in
extraction failure if latter tar was extracted before the earlier one
that created the directories. Now all relevant directories are added to
all of the tar files, ensuring they can always be independently handled
in any order.