basebackup: Ensure each tar chunk contains all relevant information
Previously the directory create operations were only part of the tar
file that contained first file/dir under the directory. This resulted in
invalid permissions being set for the directories if later tar that did
not contain the permission information for the intermediary directories
was extracted before an earlier one.

Now all relevant directories are added to all of the tar files, ensuring
they can always be handled independently in any order.