object storage: refactor iteration methods to include prefixes
Refactor the `list_iter` method into `iter_children` that yields both
objects (as `list_iter` does) and prefixes that are immediate children
of the given key. The yielded items are instances of `Child`, which is a
`namedtuple` with the `type` attribute set to `"object"` or `"prefix"`.
The `value` attribute is a `dict` for `"object"` type items, and an
`str` for `"prefix"` types.

`BaseTransfer.list_iter` and `BaseTransfer.iter_prefixes` filter over
`list_iter` items yielding only child values of the relevant type.

Backwards compatibility is maintained for `list_iter` and `list_path`.