pghoard: object storage change support
When changing object storage, pghoard would have to be restarted in order to use
the new object storage. The reason for this is that the TransferAgent would never read the
new value from the config if it already had a value stored in memory. This commit clears that value
and makes sure it gets forced to load the new value from the config.