pghoard may overwrite existing basebackup when taking a new one
PGHoard generates basebackup names locally based on timestamp and a running counter.
In scenarios when basebackup are created on a new host during the same day as an earlier basebackup
was created on an older host, the old basebackup will end up overwritten and preventing recovery from that period.

Proposed change is to change the basebackup naming from %Y-%m-%d_{COUNTER} to %Y-%m-%d_%H-%M which should solve this problem.