Make archive sync reupload first file that already exists remotely
When old node is running old pghoard version and hashes are not
available we cannot detect invalid WAL files. To opt for safety, assume
the latest file that has already been uploaded is potentially invalid
and reupload it. While this is almost always not necessary the overhead
of doing this is very small so it shouldn't be an issue.