Revert "Merge pull request #337 from pellcorp/sftp_client"
This reverts commit ae519b533086bcf7eb442d3ac6303b3693f0d445, reversing
changes made to dde7b9996ec20823cce28c384daf48e60ec42e68.

The pysftp library used for implementing the functionality seems to be
dead (last updated 3 years ago), it doesn't officially support latest
Python versions and it is not available on Fedora so after the merge
trying to build pghoard on Fedora fails and trying to use the SFTP
functionality there doesn't work.

It seems pysftp is fairly lightweight wrapper on top of paramiko and it
shouldn't be too much work to implement this without using pysftp at
all.