Fetch data from cloud storage if same file is requested again
There have been cases where a file that looks to be what PostgreSQL is
requesting already exists on disk but for some reason the file is not
acceptable and PostgreSQL keeps on requesting the same file again and
again. The version in object storage may be different and something that
PostgreSQL accepts so if we have already recently returned the requested
file from disk try object storage instead.