Allow restricting basebackups based on backup age
Often it is desirable to keep backups for certain period of time. The
backup count setting does allow emulating that to an extent but it does
not work correctly when new basebackups are created manually, in which
case the total count will be exceeded and backups are deleted before
the desired maximum age.