Tune test_pause_on_disk_full settings to fix transient failure
Processing all the WAL files can take a while because there's one
second delay each time pg_receivewal is suspended and depending on
how many times it gets suspended the test might not complete within
the previous max time. Drop segments to 15 to make the test faster
and allow it to take longer time. 15 should still be plenty to make
the disk fill up if the logic doesn't work.