Report most recently served files in pghoard state
When PostgreSQL is recovering files from archive it may not be accepting
connections so it's not possible to ask the recovery progress from PG
directly. Recovery progress in such cases is in practice the most recent
file of type "xlog" that pghoard has served. Expose this information in
the state file.