F_SETPIPE_SZ may not always work as nonroot
There are user-specific soft and hard limits totals as well, that are
not visible (and even if they are, they could potentially change as
they consist of quota left for the user from the current usage which
may change).

We try smaller and smaller sizes, up to 64kb which is the default.

ENOMEM and EBUSY are intentionally not handled.