pghoard: Break webserver parameters out to the webserver object
Move the internal configuration of OwnHTTPServer to it's own __init__
function. The use of this function looks like a misinterpretation of
the socketserver's documentation, which requires the HTTPServer __init__
not to be overridden, but does allow it to be extended.

In practice this means we call the super().__init__ method on HTTPServer
first, then initialize everything else.

This leads to a much cleaner separation of responsibilities.