Keep latest WAL file arround to enable resuming from that position
When pg_receivewal is disturbed, it will restart. The position it restarts from is figured out from the latest file in the putput directory or--if none exists--from the server position. As we delete WAL files after compression, any such disturbance could potentially lead to missing WAL segments in the backup, if there is a high enough load to advance the position further along enough to skip a WAL segment. That would result in unrestoreable backups from that position onwards until the next full basebackup is taken.

By keeping one file around, pg_receivewal can use that to start again from that position.

test_surviving_pg_receivewal_hickup tests exactly this and failed on master but not with this commit.