Workaround tar concurrency issue.
GNU tar can fail when multiple concurrent process are trying to extract
to the same directory tree. To work around this we need to retry the
failed tar command in some case.

For this, invent a new Exception "MaybeRecoverableError", and queue the
job back when it is encountered. We also have to maintain a counter for
each file, to abort after 3 times.

To test this, add a "tar_failer" command which reads it's input from
stdin and output the same message tar would do if encountering the
above-mentioned race condition.

Fixes #452.