Fix last_flushed_lsn computation
The last_flushed_lsn corresponded to the start of the wal file. This
means that we resume from the wal file which was already archived.

Instead, record the start position of the next wal file  to advance both
the replication slot and keep our internal state.

Additionally, move the completed LSN processing to it's own method, and
call it when we stop the thread (mostly useful for tests at the moment).

There is still the problem that the last_flushed_lsn is stored only in
the json state file.