Remove old backups only if there is no backup in progress
Do not clean up old backups if a backup is in progress, this might delete
delta backup files which are not referenced anywhere right now, but could
end up in manifest file of the currently running backup (race condition).
So we remove old backup with hash1, which is not referenced from anywhere
else, but currently running backup ended up with the file with the same
hash1, which it found in this old backup when backup has started, so now
we have the situation where manifest of the old backup includes reference to
this file hash, but cleanup removed it while backup was in progress.