Add metrics for running compressor threads
We want to be able to send an alert when we're about to run out of
compressor thread, since that would mean our backup process is stuck. We
send the information as a percentage so that we can set a meaningful
threshold for our alerts without the need to check each instance config.

[BF-962]