tests: test all pg versions simultaneously
This does not change much for GitHub Actions:
one dimension of the matrix is replaced by a fixture parametrization.

For local dev environment however, it makes it easier to get accurate
coverage on all version combined, without a manual merge step and
without having to think about `PG_VERSION` environment variable.

The `PG_VERSION` variable still exists and can be used to reduce the
number of tested versions, either a single version of multiple versions
separated by a comma.

Tests that don't depend on a running PostgreSQL are now run less
often since they don't depend on the parametrized fixture.