Delta backups: chunk small files
Usually when doing a backup there are many small files with the size
varying between several KBs to several MBs, which sum up to insignificant
cumulative size in comparison with the total size of DB.
It does not make much sense in this case to waste resources on calculating
hashes of those files and what is worse - spend time on uploading those to
the object storage separately (adds on latency for every file) and burn
money on every request. Delta functionality is not efficient for those.
Suggestion is to just chunk those files into archives (pretty much the same
behaviour as for local-tar) and apply delta logic for files of significant
enough size.