tests: fix flaky test_pause_on_disk_full
Fix flaky test_pause_on_disk_full test and potential other problems
with tests which require isolated directory.

Temporary directory name is not really a random combination rather the
name with a sequencial number in it, so if multiple pytests start
together, they might be reusing the same temporary directory.