walreceiver: don't split a message in two segments
If the WAL segment buffer is almost full, it's possible for the next
message to only partially fit in the buffer.

In that case, we were sending only part of the message in the current
WAL segment, and then failing because there was something left in the
buffer.

Instead, look ahead and if we know the message won't fit, split the
WAL file before adding the message.

We will still split the message (and fail) if a single message is
larger than the WAL segment size. I don't know if that can happen
and what's the correct way to handle this case.