test: add backup preservation integration test
It's not a perfect test: our setup is not well suited to testing
concurrent operations (a backup in progress and a deletion). Fixing
that is more than just adding threads, we need ways to stop between
step for instance.

To cover for the lack of concurrent execution, there is a new option
to not cancel preservation at the end of a successful restore.
It could also be useful in real scenarios, if the restoration needs
some external validation before marking it as successful.