Provide a metric measuring basebackup upload rate
If rohmu version 1.0.10 or greater is installed,
`pghoard.basebackup_bytes_uploaded` will correctly report the number of
bytes uploaded during basebackup generation.

If an earlier version of rohmu is used, the metric will still be
generated, but its value will be much smaller (but nonzero).
This may still be useful to indicate that a basebackup is in progress.