Collect upload metrics during additional basebackup phase
Where basic, piped or delta basebackup method is selected,
there is a step where data is being uploaded and the backup is
not yet available, but no activity is sent to the statsd agent.

Register an additional callback to collect the number of bytes
transferred and report it alongside the values from a previous phase.
This helps avoid unnecessary alerts due to basebackups spending time
here.