Better retry for base backup restore
Current implementation use fixed retry (6) for base backup restore.
However, on a large (2~4 TiB) database, there are more data to be
transferred and will take longer. As a result, it is more likely to hit
the limit.

This commit allows a dynamic number of retry, depends on the size of the
base backup.