Do not prefetch timeline files.
We don't have many timeline files, so it is not critical to prefetch
those, and can lead to database corruption if we try to download one,
abort the download, and try to delete it afterwards while the server has
promoted itself.

The same kind of pitfalls may exist for WAL files themselves.